--- a/pkg/webhook/advancedcronjob/mutating/advancedcronjob_create_update_handler.go
+++ b/pkg/webhook/advancedcronjob/mutating/advancedcronjob_create_update_handler.go
@@ -19,7 +19,9 @@ package mutating
 import (
 	"context"
 	"encoding/json"
+	"k8s.io/apimachinery/pkg/runtime"
 	"net/http"
+	"reflect"
 
 	appsv1alpha1 "github.com/openkruise/kruise/apis/apps/v1alpha1"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
@@ -53,11 +55,15 @@ func (h *AdvancedCronJobCreateUpdateHandler) Handle(ctx context.Context, req adm
 		return admission.Errored(http.StatusBadRequest, err)
 	}
 
+	var copy runtime.Object = obj.DeepCopy()
 	err = h.mutatingAdvancedCronJobFn(ctx, obj)
 	if err != nil {
 		return admission.Errored(http.StatusInternalServerError, err)
 	}
 
+	if reflect.DeepEqual(obj, copy) {
+		return admission.Allowed("")
+	}
 	marshalled, err := json.Marshal(obj)
 	if err != nil {
 		return admission.Errored(http.StatusInternalServerError, err)