--- a/pkg/webhook/broadcastjob/mutating/broadcastjob_create_update_handler.go
+++ b/pkg/webhook/broadcastjob/mutating/broadcastjob_create_update_handler.go
@@ -19,7 +19,9 @@ package mutating
 import (
 	"context"
 	"encoding/json"
+	"k8s.io/apimachinery/pkg/runtime"
 	"net/http"
+	"reflect"
 
 	appsv1alpha1 "github.com/openkruise/kruise/apis/apps/v1alpha1"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
@@ -52,10 +54,13 @@ func (h *BroadcastJobCreateUpdateHandler) Handle(ctx context.Context, req admiss
 	if err != nil {
 		return admission.Errored(http.StatusBadRequest, err)
 	}
-
+	var copy runtime.Object = obj.DeepCopy()
 	err = h.mutatingBroadcastJobFn(ctx, obj)
 	if err != nil {
 		return admission.Errored(http.StatusInternalServerError, err)
+	}
+	if reflect.DeepEqual(obj, copy) {
+		return admission.Allowed("")
 	}
 
 	marshalled, err := json.Marshal(obj)