--- a/pkg/webhook/daemonset/mutating/daemonset_create_update_handler.go
+++ b/pkg/webhook/daemonset/mutating/daemonset_create_update_handler.go
@@ -19,7 +19,9 @@ package mutating
 import (
 	"context"
 	"encoding/json"
+	"k8s.io/apimachinery/pkg/runtime"
 	"net/http"
+	"reflect"
 
 	appsv1alpha1 "github.com/openkruise/kruise/apis/apps/v1alpha1"
 	"github.com/openkruise/kruise/pkg/util"
@@ -49,9 +51,11 @@ func (h *DaemonSetCreateUpdateHandler) Handle(ctx context.Context, req admission
 	if err != nil {
 		return admission.Errored(http.StatusBadRequest, err)
 	}
-
+	var copy runtime.Object = obj.DeepCopy()
 	appsv1alpha1.SetDefaultsDaemonSet(obj)
-
+	if reflect.DeepEqual(obj, copy) {
+		return admission.Allowed("")
+	}
 	marshalled, err := json.Marshal(obj)
 	if err != nil {
 		return admission.Errored(http.StatusInternalServerError, err)