--- a/pkg/webhook/imagepulljob/mutating/imagepulljob_create_update_handler.go
+++ b/pkg/webhook/imagepulljob/mutating/imagepulljob_create_update_handler.go
@@ -20,6 +20,9 @@ import (
 	"context"
 	"encoding/json"
 	"net/http"
+	"reflect"
+
+	"k8s.io/apimachinery/pkg/runtime"
 
 	appsv1alpha1 "github.com/openkruise/kruise/apis/apps/v1alpha1"
 	"github.com/openkruise/kruise/pkg/util"
@@ -42,9 +45,11 @@ func (h *ImagePullJobCreateUpdateHandler) Handle(ctx context.Context, req admiss
 	if err != nil {
 		return admission.Errored(http.StatusBadRequest, err)
 	}
-
+	var copy runtime.Object = obj.DeepCopy()
 	appsv1alpha1.SetDefaultsImagePullJob(obj)
-
+	if reflect.DeepEqual(obj, copy) {
+		return admission.Allowed("")
+	}
 	marshalled, err := json.Marshal(obj)
 	if err != nil {
 		return admission.Errored(http.StatusInternalServerError, err)