--- a/pkg/webhook/nodeimage/mutating/nodeimage_create_update_handler.go
+++ b/pkg/webhook/nodeimage/mutating/nodeimage_create_update_handler.go
@@ -19,7 +19,9 @@ package mutating
 import (
 	"context"
 	"encoding/json"
+	"k8s.io/apimachinery/pkg/runtime"
 	"net/http"
+	"reflect"
 
 	appsv1alpha1 "github.com/openkruise/kruise/apis/apps/v1alpha1"
 	"github.com/openkruise/kruise/pkg/util"
@@ -43,7 +45,7 @@ func (h *NodeImageCreateUpdateHandler) Handle(ctx context.Context, req admission
 	if err != nil {
 		return admission.Errored(http.StatusBadRequest, err)
 	}
-
+	var copy runtime.Object = obj.DeepCopy()
 	// Set defaults
 	appsv1alpha1.SetDefaultsNodeImage(obj)
 
@@ -56,7 +58,9 @@ func (h *NodeImageCreateUpdateHandler) Handle(ctx context.Context, req admission
 		utilimagejob.SortStatusImageTags(&imageStatus)
 		obj.Status.ImageStatuses[name] = imageStatus
 	}
-
+	if reflect.DeepEqual(obj, copy) {
+		return admission.Allowed("")
+	}
 	marshalled, err := json.Marshal(obj)
 	if err != nil {
 		return admission.Errored(http.StatusInternalServerError, err)