--- a/pkg/webhook/sidecarset/mutating/sidecarset_create_update_handler.go
+++ b/pkg/webhook/sidecarset/mutating/sidecarset_create_update_handler.go
@@ -19,7 +19,9 @@ package mutating
 import (
 	"context"
 	"encoding/json"
+	"k8s.io/apimachinery/pkg/runtime"
 	"net/http"
+	"reflect"
 
 	appsv1alpha1 "github.com/openkruise/kruise/apis/apps/v1alpha1"
 	"github.com/openkruise/kruise/pkg/control/sidecarcontrol"
@@ -72,7 +74,7 @@ func (h *SidecarSetCreateHandler) Handle(ctx context.Context, req admission.Requ
 	if err != nil {
 		return admission.Errored(http.StatusBadRequest, err)
 	}
-
+	var copy runtime.Object = obj.DeepCopy()
 	switch req.AdmissionRequest.Operation {
 	case v1beta1.Create, v1beta1.Update:
 		appsv1alpha1.SetDefaultsSidecarSet(obj)
@@ -81,7 +83,9 @@ func (h *SidecarSetCreateHandler) Handle(ctx context.Context, req admission.Requ
 		}
 	}
 	klog.V(4).Infof("sidecarset after mutating: %v", util.DumpJSON(obj))
-
+	if reflect.DeepEqual(obj, copy) {
+		return admission.Allowed("")
+	}
 	marshalled, err := json.Marshal(obj)
 	if err != nil {
 		return admission.Errored(http.StatusInternalServerError, err)